difftest: always check all instruction commits
We allow invalid instr[0] with valid instr[1] by looping over all
instruction commits.

In out-of-order designs, instructions commits generally happen
in-order and the head commit slow will be the first instruction.
This is usually implemented with a circular buffer.

However, in in-order multi-issue designs, we don't require the
design to align the commits for difftest.